Avoid importing auth if not needed.
