+++ tiled/server/utils.py
@@ -1,6 +1,20 @@
 import contextlib
 import time
 
+ACCESS_TOKEN_COOKIE_NAME = "tiled_access_token"
+REFRESH_TOKEN_COOKIE_NAME = "tiled_refresh_token"
+API_KEY_COOKIE_NAME = "tiled_api_key"
+API_KEY_HEADER_NAME = "x-tiled-api-key"
+API_KEY_QUERY_PARAMETER = "api_key"
+CSRF_COOKIE_NAME = "tiled_csrf"
+
+
+def get_authenticators():
+    raise NotImplementedError(
+        "This should be overridden via dependency_overrides. "
+        "See tiled.server.app.build_app()."
+    )
+
 
 @contextlib.contextmanager
 def record_timing(metrics, key):
